// src/api/__generated__/story/model/listDramasParams.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Mini Drama Service API
 * Mini Drama Platform
 * OpenAPI spec version: v1.0
 */
import type { ListDramasStatus } from './listDramasStatus';

export type ListDramasParams = {
  /**
   * 短剧状态
   */
  status?: ListDramasStatus;
  /**
   * @minimum -1
   */
  mark?: number;
  /**
   * @minimum 1
   * @maximum 100
   */
  pageSize?: number;
};